[common/flights] add flights for query&store action
