+++ common/flights/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "common-flights"
+version = "0.1.0"
+authors = ["Datafuse Authors <opensource@datafuselabs.com>"]
+license = "Apache-2.0"
+publish = false
+edition = "2018"
+
+[dependencies] # In alphabetical order
+# Workspace dependencies
+common-arrow = {path = "../arrow"}
+common-planners = {path = "../planners"}
+
+# Github dependencies
+
+# Crates.io dependencies
+anyhow = "1.0.40"
+prost = "0.7"
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+tonic = "0.4"
+
+[dev-dependencies]
+pretty_assertions = "0.7"
+
+[build-dependencies]
+tonic-build = "0.4"
