+++ common/flights/build.rs
@@ -1,0 +1,28 @@
+// Copyright 2020-2021 The Datafuse Authors.
+//
+// SPDX-License-Identifier: Apache-2.0.
+
+use std::env;
+use std::path::Path;
+
+fn main() {
+    build_proto();
+}
+
+fn build_proto() {
+    let manifest_dir =
+        env::var("CARGO_MANIFEST_DIR").expect("CARGO_MANIFEST_DIR env variable unset");
+
+    let proto_dir = Path::new(&manifest_dir).join("../../proto");
+    let protos = [
+        &Path::new(&proto_dir).join(Path::new("queryflight.proto")),
+        &Path::new(&proto_dir).join(Path::new("storeflight.proto")),
+    ];
+
+    for proto in protos.iter() {
+        println!("cargo:rerun-if-changed={}", proto.to_str().unwrap());
+    }
+    tonic_build::configure()
+        .compile(&protos, &[&proto_dir])
+        .unwrap();
+}
