+++ fusequery/query/Cargo.toml
@@ -16,6 +16,7 @@ path = "src/bin/fuse-query.rs"
 common-arrow = {path = "../../common/arrow"}
 common-datablocks = {path = "../../common/datablocks"}
 common-datavalues = {path = "../../common/datavalues"}
+common-flights = {path = "../../common/flights"}
 common-functions = {path = "../../common/functions"}
 common-infallible = {path = "../../common/infallible"}
 common-planners = {path = "../../common/planners"}
@@ -42,8 +43,8 @@ num_cpus = "1.0"
 paste = "^1.0"
 prost = "0.7"
 rand = "0.8.3"
-serde_json = "1.0"
 serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
 sqlparser = "0.9"
 structopt = "0.3"
 threadpool = "1.8.1"
@@ -55,6 +56,3 @@ uuid = { version = "0.8", features = ["serde", "v4"] }
 
 [dev-dependencies]
 pretty_assertions = "0.7"
-
-[build-dependencies]
-tonic-build = "0.4"
