+++ fusequery/query/build.rs
@@ -16,7 +16,6 @@ fn main() {
     }
 
     create_version_info();
-    build_proto();
 }
 
 fn create_version_info() {
@@ -55,19 +54,4 @@ fn commit_date() -> Option<String> {
         .output()
         .ok()
         .and_then(|r| String::from_utf8(r.stdout).ok())
-}
-
-fn build_proto() {
-    let manifest_dir =
-        env::var("CARGO_MANIFEST_DIR").expect("CARGO_MANIFEST_DIR env variable unset");
-
-    let proto_dir = Path::new(&manifest_dir).join("../../proto");
-    let protos = [&Path::new(&proto_dir).join(Path::new("queryflight.proto"))];
-
-    for proto in protos.iter() {
-        println!("cargo:rerun-if-changed={}", proto.to_str().unwrap());
-    }
-    tonic_build::configure()
-        .compile(&protos, &[&proto_dir])
-        .unwrap();
 }
