+++ fusequery/query/src/api/rpc/mod.rs
@@ -8,11 +8,9 @@ mod flight_service_test;
 #[macro_use]
 mod macros;
 
-mod flight_action;
 mod flight_client;
 mod flight_service;
 mod metrics;
 
-pub use flight_action::{DoActionAction, DoGetAction, ExecutePlanAction, FetchPartitionAction};
 pub use flight_client::FlightClient;
 pub use flight_service::{FlightService, FlightStream};
